requirements.txt
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
expiringdict==1.2.2
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
launchdarkly-eventsource==1.3.0
launchdarkly-server-sdk==9.11.1
MarkupSafe==3.0.2
pyRFC3339==2.0.1
python-dotenv==1.1.1
semver==3.0.4
urllib3==2.5.0
Werkzeug==3.1.3
